# test-tools/swagger-src/backend_swagger20.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "",
    "version" : "1.0.14 (20180124)",
    "title" : "BOCHK FinTech Hackathon 2018 API (Backend)",
    "contact" : {
      "url" : "http://www.bochkhackathon.com/",
      "email" : "it_innovation_lab@bochk.com"
    }
  },
  "host" : "bocapis.mybluemix.net",
  "basePath" : "/api",
  "schemes" : [ "https" ],
  "consumes" : [ "application/json" ],
  "produces" : [ "application/json" ],
  "paths" : {
...
    "/services/appointments/{phone_no}" : {
      "get" : {
        "tags" : [ "Appointments" ],
        "summary" : "List appointments",
        "parameters" : [ {
          "name" : "ibm-app-user",
          "in" : "header",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Status 200",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/appointment"
              }
            }
          },
          "404" : {
            "description" : "Status 404"
          }
        }
      },
      "post" : {
        "tags" : [ "Appointments" ],
        "summary" : "Create appointment",
        "consumes" : [ ],
        "parameters" : [ {
          "name" : "ibm-app-user",
          "in" : "header",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "body",
          "in" : "body",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/appointment_creation_request"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Appointment created",
            "schema" : {
              "$ref" : "#/definitions/appointment"
            }
          },
          "400" : {
            "description" : "Status 400"
          }
        }
      },
      "parameters" : [ {
        "name" : "phone_no",
        "in" : "path",
        "required" : true,
        "type" : "string"
      } ]
    },
    "/services/appointments/{phone_no}/{appointment_id}" : {
      "delete" : {
        "tags" : [ "Appointments" ],
        "summary" : "Cancel appointment",
        "parameters" : [ {
          "name" : "ibm-app-user",
          "in" : "header",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Status 200"
          },
          "400" : {
            "description" : "Status 400",
            "schema" : {
              "$ref" : "#/definitions/error_response"
            }
          },
          "404" : {
            "description" : "Status 404",
            "schema" : {
              "$ref" : "#/definitions/error_response"
            }
          }
        }
      },
      "parameters" : [ {
        "name" : "phone_no",
        "in" : "path",
        "required" : true,
        "type" : "string"
      }, {
        "name" : "appointment_id",
        "in" : "path",
        "required" : true,
        "type" : "string"
      } ]
    },
...
  },
  "securityDefinitions" : {
    "Basic authentication" : {
      "type" : "basic"
    }
  },
  "definitions" : {
...
    "appointment_creation_request" : {
      "type" : "object",
      "required" : [ "branch_code", "datetime", "requested_service_category" ],
      "properties" : {
        "branch_code" : {
          "type" : "string"
        },
        "datetime" : {
          "type" : "string",
          "format" : "date-time"
        },
        "requested_service_category" : {
          "type" : "string",
          "enum" : [ "WEALTH_MANAGEMENT", "PRIVATE_BANKING", "CORP_BANKING", "RMB_SERVICES", "CROSS_BORDER_SERVICES", "INVESTMENT", "MORTGAGE", "LOAN", "OTHERS" ]
        },
        "requested_service_detail" : {
          "type" : "string"
        }
      }
    },
...
    "appointment" : {
      "type" : "object",
      "required" : [ "appointment_id", "branch_code", "datetime", "phone_no", "requested_service_category" ],
      "properties" : {
        "appointment_id" : {
          "type" : "string"
        },
        "phone_no" : {
          "type" : "string"
        },
        "branch_code" : {
          "type" : "string"
        },
        "datetime" : {
          "type" : "string",
          "format" : "date-time"
        },
        "requested_service_category" : {
          "type" : "string",
          "enum" : [ "WEALTH_MANAGEMENT", "PRIVATE_BANKING", "CORP_BANKING", "RMB_SERVICES", "CROSS_BORDER_SERVICES", "INVESTMENT", "MORTGAGE", "LOAN", "OTHERS" ]
        },
        "requested_service_detail" : {
          "type" : "string"
        }
      }
    },
    "error_response" : {
      "type" : "object",
      "required" : [ "code", "message" ],
      "properties" : {
        "code" : {
          "type" : "string"
        },
        "message" : {
          "type" : "string"
        }
      }
    }
  }
}